requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-extensions==3.2.3
django-sslserver==0.22
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
gunicorn==23.0.0
inflection==0.5.1
MarkupSafe==3.0.2
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
Werkzeug==3.1.3
